Revert "Revert "ExtensionDEscriptorParser""
This reverts commit cb706e49cb88518314e6f6e3d34d586f15b28b93.
