+++ com.sap.cloud.lm.sl.mta/src/main/java/com/sap/cloud/lm/sl/mta/parsers/v2/ExtensionDescriptorParser.java
@@ -2,30 +2,48 @@ package com.sap.cloud.lm.sl.mta.parsers.v2;
 
 import static com.sap.cloud.lm.sl.mta.handlers.v2.Schemas.MTAEXT;
 
+import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
+import java.util.Set;
 
 import com.sap.cloud.lm.sl.common.ParsingException;
 import com.sap.cloud.lm.sl.common.util.ListUtil;
 import com.sap.cloud.lm.sl.mta.model.v2.ExtensionDescriptor;
+import com.sap.cloud.lm.sl.mta.model.v2.ExtensionDescriptor.Builder;
 import com.sap.cloud.lm.sl.mta.model.v2.ExtensionModule;
 import com.sap.cloud.lm.sl.mta.model.v2.ExtensionResource;
-import com.sap.cloud.lm.sl.mta.model.v2.ExtensionDescriptor.Builder;
+import com.sap.cloud.lm.sl.mta.parsers.ListParser;
+import com.sap.cloud.lm.sl.mta.parsers.ModelParser;
 import com.sap.cloud.lm.sl.mta.schema.MapElement;
 
-public class ExtensionDescriptorParser extends com.sap.cloud.lm.sl.mta.parsers.v1.ExtensionDescriptorParser {
+public class ExtensionDescriptorParser extends ModelParser<ExtensionDescriptor> {
 
+    protected static final String PROCESSED_OBJECT_NAME = "MTA extension descriptor";
+    public static final String ID = "ID";
+    public static final String EXT_DESCRIPTION = "ext_description";
+    public static final String EXTENDS = "extends";
+    public static final String EXT_PROVIDER = "ext_provider";
     public static final String DESCRIPTION = "description";
     public static final String PROVIDER = "provider";
+    public static final String TARGET_PLATFORMS = "target-platforms";
+    public static final String MODULES = "modules";
+    public static final String RESOURCES = "resources";
+    public static final String PROPERTIES = "properties";
+    public static final String SCHEMA_VERSION = "_schema-version";
     public static final String PARAMETERS = "parameters";
     public static final String DEPLOY_TARGETS = "targets";
 
+    protected final Set<String> usedModuleNames = new HashSet<>();
+    protected final Set<String> usedProvidedDependencyNames = new HashSet<>();
+    protected final Set<String> usedResourceNames = new HashSet<>();
+
     public ExtensionDescriptorParser(Map<String, Object> source) {
-        super(MTAEXT, source);
+        this(MTAEXT, source);
     }
 
     protected ExtensionDescriptorParser(MapElement schema, Map<String, Object> source) {
-        super(schema, source);
+        super(PROCESSED_OBJECT_NAME, schema, source);
     }
 
     @Override
@@ -51,7 +69,6 @@ public class ExtensionDescriptorParser extends com.sap.cloud.lm.sl.mta.parsers.v
      * @deprecated Use {@link #getDescription()} instead.
      */
     @Deprecated
-    @Override
     protected String getExtensionDescription() {
         return getDescription();
     }
@@ -60,7 +77,6 @@ public class ExtensionDescriptorParser extends com.sap.cloud.lm.sl.mta.parsers.v
      * @deprecated Use {@link #getProvider()} instead.
      */
     @Deprecated
-    @Override
     protected String getExtensionProvider() {
         return getProvider();
     }
@@ -81,6 +97,37 @@ public class ExtensionDescriptorParser extends com.sap.cloud.lm.sl.mta.parsers.v
         return getStringElement(EXT_PROVIDER);
     }
 
+    protected String getSchemaVersion() {
+        return getSchemaVersion(SCHEMA_VERSION);
+    }
+
+    protected String getId() {
+        return getStringElement(ID);
+    }
+
+    protected String getParentId() {
+        return getStringElement(EXTENDS);
+    }
+
+    protected List<String> getTargetPLatforms() {
+        return getListElement(TARGET_PLATFORMS);
+    }
+
+    protected List<ExtensionModule> getModules2() {
+        return ListUtil.cast(getModules());
+    }
+
+    protected List<ExtensionModule> getModules() {
+        return getListElement(MODULES, new ListParser<ExtensionModule>() {
+            @Override
+            protected ExtensionModule parseItem(Map<String, Object> map) {
+                return getModuleParser(map).setUsedProvidedDependencyNames(usedProvidedDependencyNames)
+                    .setUsedValues(usedModuleNames)
+                    .parse();
+            }
+        });
+    }
+
     protected List<String> getDeployTargets() {
         return getListElement(DEPLOY_TARGETS);
     }
@@ -90,19 +137,23 @@ public class ExtensionDescriptorParser extends com.sap.cloud.lm.sl.mta.parsers.v
     }
 
     protected List<ExtensionResource> getResources2() {
-        return ListUtil.cast(getResources1());
+        return ListUtil.cast(getResources());
     }
 
-    @Override
+    protected List<ExtensionResource> getResources() {
+        return getListElement(RESOURCES, new ListParser<ExtensionResource>() {
+            @Override
+            protected ExtensionResource parseItem(Map<String, Object> map) {
+                return getResourceParser(map).setUsedValues(usedResourceNames)
+                    .parse();
+            }
+        });
+    }
+
     protected ExtensionResourceParser getResourceParser(Map<String, Object> source) {
         return new ExtensionResourceParser(source); // v2
     }
 
-    protected List<ExtensionModule> getModules2() {
-        return ListUtil.cast(getModules1());
-    }
-
-    @Override
     protected ExtensionModuleParser getModuleParser(Map<String, Object> source) {
         return new ExtensionModuleParser(source); // v2
     }
